requirements.txt
flask-socketio
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
pandas==0.25.3
Pillow==6.1.0
nltk==3.4.4
pyspellchecker
requests==2.21.0
scapy==2.4.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.1
seaborn==0.9.0
SpeechRecognition==3.8.1
tensorflow
tflearn
